+++ src/components/Synthetics/TradeBox/TradeBox.scss
@@ -126,6 +126,11 @@
   }
 }
 
+.LiqPrice-tooltip-warning {
+  color: rgb(243, 181, 12);
+  text-decoration-color: rgb(243, 181, 12);
+}
+
 @media (max-width: 900px) {
   .SwapBox-content {
     flex-wrap: wrap;
